Address kotlinx.serialization in interplay with arc
Ktor was already upgraded to version 3.x in arc - and this leads to incompatibilities with different versions of kotlinx.serialization being used in lmos-runtime and in the arc libraries pulled in by the lmos-runtime.

To address this, the ktor version in the runtime is also upgraded to 3.x, and the kotlinx.serialization version is increased to 1.8.0.

As spring boot's gradle plugin manages the kotlinx.serialization version, we need to use a dependencyManagement block in the two modules pulling in that plugin to actually increase the version of kotlinx.serialization.